client/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mypy==1.3.0
mypy-extensions==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tomli==2.0.1
types-python-dateutil==2.8.19.13
types-requests==2.31.0.0
types-urllib3==1.26.25.13
typing-extensions==4.6.2
urllib3==2.0.2
